fix(paper-trading): wire signal approval to create positions and add crypto support
- Fix critical bug where approving signals marked them as executed but
  didn't create positions (was a TODO comment)
- Add crypto price support to markets router using Binance API
- Support fractional quantities for crypto positions (e.g., 0.238 BTC)
- Update API schemas to use float for quantity instead of int
- Fix documentation to match actual API endpoints

+++ api/schemas.py
@@ -57,7 +57,7 @@ class PositionResponse(BaseModel):
     symbol: str
     market: str
     direction: str
-    quantity: int
+    quantity: float  # Float to support fractional crypto quantities
     entry_price: float
     current_price: float
     stop_loss: float
@@ -88,7 +88,7 @@ class TradeResponse(BaseModel):
     position_id: Optional[int]
     symbol: str
     trade_type: str
-    quantity: int
+    quantity: float  # Float to support fractional crypto quantities
     price: float
     value: float
     commission: float
